use a more robust check for class membership
I got some warnings running susie (I'd be happy to generate a reproducible example if you'd like), so I thought I'd make a pull request fixing the problem. `inherits(foo,"bar")` is more or less equivalent to `any(class(foo)=="bar")`.  In general it's not a good idea to assume the return value of `class` is scalar.  The warning was :
```
Warning message:
In if (class(res) == "susie") res = res$alpha :
  the condition has length > 1 and only the first element will be used
```
and comes from setting 
```
_R_CHECK_LENGTH_1_CONDITION_=verbose
_R_CHECK_LENGTH_1_LOGIC2_=verbose
```
in `.Renviron`