speed-up for susie_plot
The susie_plot function is slow when genomic positions are used. It also creates spurious points on the plot at genomic positions that are supposed to be empty. This change removes the spurious points  by using NA instead of min(p) to initialize empty positions. This also results in a partial speed-up of the function. The rest of the speed-up is achieved by keeping model$alpha unchanged when genomic positions are specified. This means that model$alpha only non-empty positions, reducing runtime for the in_CS call. The output of the in_CS call is then matched back to non-empty positions